move sui sdk tests
