+++ libs/coin-modules/coin-sui/src/network/sdk.integration.test.ts
@@ -1,6 +1,7 @@
 import BigNumber from "bignumber.js";
-import { getOperations } from "./sdk";
+import { createTransaction, getOperations } from "./sdk";
 import type { Operation } from "@ledgerhq/types-live";
+import { getAccount } from "./sdk";
 
 describe("getOperations", () => {
   describe("Account 0x33444cf803c690db96527cec67e3c9ab512596f4ba2d4eace43f0b4f716e0164", () => {
@@ -71,5 +72,29 @@ describe("getOperations", () => {
         });
       });
     });
+  });
+});
+
+describe("getBalance", () => {
+  test("getAccount should return account balance", async () => {
+    const address = "0x33444cf803c690db96527cec67e3c9ab512596f4ba2d4eace43f0b4f716e0164";
+    const balance = await getAccount(address);
+    expect(balance).toHaveProperty("blockHeight");
+    expect(balance).toHaveProperty("balance");
+  });
+});
+
+describe("createTransaction", () => {
+  test("createTransaction should build a transaction", async () => {
+    const address = "0x6e143fe0a8ca010a86580dafac44298e5b1b7d73efc345356a59a15f0d7824f0";
+    const transaction = {
+      mode: "sent",
+      family: "sui" as const,
+      amount: new BigNumber(100),
+      recipient: "0x33444cf803c690db96527cec67e3c9ab512596f4ba2d4eace43f0b4f716e0164",
+      errors: {},
+    };
+    const tx = await createTransaction(address, transaction);
+    expect(tx).toBeInstanceOf(Uint8Array);
   });
 });
