+++ libs/coin-modules/coin-sui/src/network/sdk.test.ts
@@ -1,6 +1,5 @@
 import { TransactionBlockData, SuiTransactionBlockResponse } from "@mysten/sui/client";
 import {
-  getAccount,
   getOperationType,
   getOperationSenders,
   getOperationRecipients,
@@ -8,9 +7,7 @@ import {
   getOperationFee,
   getOperationDate,
   transactionToOperation,
-  getOperations,
   paymentInfo,
-  createTransaction,
 } from "./sdk";
 
 import { BigNumber } from "bignumber.js";
@@ -109,13 +106,6 @@ const mockTransaction = {
 };
 
 describe("SDK Functions", () => {
-  test("getAccount should return account balance", async () => {
-    const address = "0x33444cf803c690db96527cec67e3c9ab512596f4ba2d4eace43f0b4f716e0164";
-    const balance = await getAccount(address);
-    expect(balance).toHaveProperty("blockHeight");
-    expect(balance).toHaveProperty("balance");
-  });
-
   test("getOperationType should return IN for incoming tx", () => {
     const address = "0x33444cf803c690db96527cec67e3c9ab512596f4ba2d4eace43f0b4f716e0164";
     expect(
@@ -166,13 +156,6 @@ describe("SDK Functions", () => {
     expect(operation).toHaveProperty("accountId", accountId);
   });
 
-  test("getOperations should fetch operations", async () => {
-    const accountId = "mockAccountId";
-    const addr = "0x33444cf803c690db96527cec67e3c9ab512596f4ba2d4eace43f0b4f716e0164";
-    const operations = await getOperations(accountId, addr);
-    expect(Array.isArray(operations)).toBe(true);
-  });
-
   test("paymentInfo should return gas budget and fees", async () => {
     const sender = "0x6e143fe0a8ca010a86580dafac44298e5b1b7d73efc345356a59a15f0d7824f0";
     const fakeTransaction = {
@@ -186,18 +169,5 @@ describe("SDK Functions", () => {
     expect(info).toHaveProperty("gasBudget");
     expect(info).toHaveProperty("totalGasUsed");
     expect(info).toHaveProperty("fees");
-  });
-
-  test("createTransaction should build a transaction", async () => {
-    const address = "0x6e143fe0a8ca010a86580dafac44298e5b1b7d73efc345356a59a15f0d7824f0";
-    const transaction = {
-      mode: "sent",
-      family: "sui" as const,
-      amount: new BigNumber(100),
-      recipient: "0x33444cf803c690db96527cec67e3c9ab512596f4ba2d4eace43f0b4f716e0164",
-      errors: {},
-    };
-    const tx = await createTransaction(address, transaction);
-    expect(tx).toBeInstanceOf(Uint8Array);
   });
 });
